KB-11836 | DEV | BE | Enablement for MDO Report via new hierarchy
1. Removed the WarehouseDatasource configuration.
2. Moved the code to new class files.
