+++ src/main/resources/application.properties
@@ -22,14 +22,6 @@ wingspan.datasource.driver-class-name=org.postgresql.Driver
 wingspan.jpa.properties.hibernate.format_sql=true
 wingspan.jpa.show-sql=false
 
-# warehouse datasource (PostgreSQL Database)
-warehouse.datasource.jdbc-url=jdbc:postgresql://postgresIP:5432/warehouse
-warehouse.datasource.username=postgres
-warehouse.datasource.password=password123
-warehouse.datasource.driver-class-name=org.postgresql.Driver
-warehouse.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
-warehouse.jpa.properties.hibernate.format_sql=true
-warehouse.jpa.show-sql=false
 
 #Workflow - Service details
 wf.service.host=http://localhost:5099/
